requirements.txt
pytest==7.4.2
selenium==4.13.0
webdriver-manager==4.0.1
pytest-html==4.0.2
allure-pytest==2.15.0
python-dotenv==1.2.1
